refactor(Commit, CommitEditor, Readme, ReadmeEditor): containerize!
Add selections for all needed props!
Change name from `Commit` to `CommitEditor` and `CommitHistory` to Commit`, `Readme` to `ReadmeEditor`, and `ReadmeHistory` to `Readme`
- containerized `ReadmeEditor` and `Readme`
