+++ app/components/workbench/Readme.tsx
@@ -2,21 +2,20 @@ import * as React from 'react'
 import { connect } from 'react-redux'
 
 import Store from '../../models/store'
+import { selectHistoryDatasetRef } from '../../selections'
 
-export interface ReadmeHistoryProps {
-  peername: string
-  name: string
-  path: string
+export interface ReadmeProps {
+  datasetRef: string
 }
 
-export const ReadmeHistoryComponent: React.FunctionComponent<ReadmeHistoryProps> = (props) => {
-  const { peername, name, path } = props
+export const ReadmeComponent: React.FunctionComponent<ReadmeProps> = (props) => {
+  const { datasetRef } = props
   const [hasReadme, setHasReadme] = React.useState(true)
   const refWithCallback = () => {
     const ref = React.useRef<HTMLDivElement>(null)
     const setRef = React.useCallback((el: HTMLDivElement) => {
       if (el !== null) {
-        fetch(`http://localhost:2503/render/${peername}/${name}/at/${path}`)
+        fetch(`http://localhost:2503/render/${datasetRef}`)
           .then(async (res) => {
             return res.text()
           })
@@ -25,7 +24,7 @@ export const ReadmeHistoryComponent: React.FunctionComponent<ReadmeHistoryProps>
             el.innerHTML = render
           })
       }
-    }, [peername, name, path])
+    }, [datasetRef])
     ref.current = setRef
     return [setRef]
   }
@@ -46,17 +45,11 @@ export const ReadmeHistoryComponent: React.FunctionComponent<ReadmeHistoryProps>
 }
 
 const mapStateToProps = (state: Store) => {
-  const { selections } = state
-
-  const { peername, name, commit } = selections
-
   // get data for the currently selected component
   return {
-    peername,
-    path: commit,
-    name
+    datasetRef: selectHistoryDatasetRef(state)
   }
 }
 
 // TODO (b5) - this doesn't need to be a container at all
-export default connect(mapStateToProps)(ReadmeHistoryComponent)
+export default connect(mapStateToProps)(ReadmeComponent)
