+++ app/selections.ts
@@ -1,9 +1,11 @@
-import Dataset from "./models/dataset"
+import Dataset, { Commit } from "./models/dataset"
 import cloneDeep from 'clone-deep'
 
 import Store, { CommitDetails, Status } from "./models/store"
 
-export function selectDatasetFromMutations (state: Store): Dataset {
+// combines working dataset and mutations dataset to return most
+// up-to-date version of the edited dataset
+export function selectMutationsDataset (state: Store): Dataset {
   const { mutations } = state
   const mutationsDataset = mutations.dataset.value
 
@@ -12,12 +14,66 @@ export function selectDatasetFromMutations (state: Store): Dataset {
   return d
 }
 
+export function selectHistoryCommit (state: Store): Commit | undefined {
+  return selectHistoryDataset(state).commit
+}
+
+// returns a dataset that only contains components
+export function selectHistoryDataset (state: Store): Dataset {
+  return datasetFromCommitDetails(state.commitDetails)
+}
+
+export function selectHistoryDatasetRef (state: Store): string {
+  return `${state.commitDetails.peername}/${state.commitDetails.name}/at${state.commitDetails.path}`
+}
+
+export function selectHistoryStatus (state: Store): Status {
+  return state.commitDetails.status
+}
+
+export function selectIsCommiting (state: Store): boolean {
+  return state.mutations.save.isLoading
+}
+
+export function selectIsLinked (state: Store): boolean {
+  return !!state.workingDataset.fsiPath && state.workingDataset.fsiPath !== ''
+}
+
+export function selectMutationsCommit (state: Store): Commit {
+  return state.mutations.save.value
+}
+
+export function selectStatusFromMutations (state: Store): Status {
+  const { workingDataset, mutations } = state
+  const mutationsStatus = mutations.status.value
+
+  return { ...workingDataset.status, ...mutationsStatus }
+}
+
+// returns a dataset that only contains components
 export function selectWorkingDataset (state: Store): Dataset {
   return datasetFromCommitDetails(state.workingDataset)
 }
 
-export function selectHistoryDataset (state: Store): Dataset {
-  return datasetFromCommitDetails(state.commitDetails)
+export function selectWorkingDatasetIsLoading (state: Store): boolean {
+  return state.workingDataset.isLoading
+}
+
+export function selectWorkingDatasetName (state: Store): string {
+  return state.workingDataset.name
+}
+
+export function selectWorkingDatasetPeername (state: Store): string {
+  return state.workingDataset.peername
+}
+
+// returns username/datasetname
+export function selectWorkingDatasetRef (state: Store): string {
+  return `${state.workingDataset.peername}/${state.workingDataset.name}`
+}
+
+export function selectWorkingStatus (state: Store): Status {
+  return state.workingDataset.status
 }
 
 function datasetFromCommitDetails (commitDetails: CommitDetails): Dataset {
@@ -31,19 +87,4 @@ function datasetFromCommitDetails (commitDetails: CommitDetails): Dataset {
     }
   })
   return d
-}
-
-export function selectStatusFromMutations (state: Store): Status {
-  const { workingDataset, mutations } = state
-  const mutationsStatus = mutations.status.value
-
-  return { ...workingDataset.status, ...mutationsStatus }
-}
-
-export function selectWorkingStatus (state: Store): Status {
-  return state.workingDataset.status
-}
-
-export function selectHistoryStatus (state: Store): Status {
-  return state.commitDetails.status
 }
